All info windows refactored as InfoCards.
This is a cleaner approach and while it does not reduce the number of
lines of code it makes it more readable. I'm using a prop to pass a
component which I think is like a render prop but still not sure what
those are. Fragments made one of the windows much easier vis. styling
I think if I were to go further on this the next step would be to create
a HOC that returned the InfoCards and maybe add the hidden state to each
instead of housing it in parent, not sure why parent cares.